Split level.py into multiple files, defined Wall class
Split the level.py file into level.py, cell.py, and 1stpcg.py

level.py contains the Level class.  The init function for this class
was refactored.  The init function was changed to create a level made
completely of Walls instead of completely of blank space. Added header
comments and a brief class description.

cell.py contains the Cell class and all of its subclasses, currently just
the Wall class.  Added header comments and brief class descriptions.

1stpcg.py serves as the executable for the program.  Currently prompts the
user for a width and height, generates a level of that width and height,
and prints the level layout.