Added Randomly Generated Start Position
Start position is now randomly generated upon level creation.  A start
side can be provided, in which case the start position is somewhere
random on that side, or else one is chosen randomly.

Currently, the start position is generated correctly, and the Char sprite's
start position is correctly based on the start position.  The camera
jumps to the character, but sometimes positions the camera such that area
outside of the level is visible.

Committing to back up before changing camera movement code.