Added Finish Cells
Finish cells are randomly generated and appear on any outside wall
except the one the start cell is on. If the finish cell is on a wall
perpendicular to the one that the start cell is on, it is generated
on the far half of the wall.

The wall, start, and finish sprites are now added to seperate groups
for rendering and collision.

When the Char sprite is within the finish sprite, a new level is
generated.